Adding content
I started by following a mixture of videos from Garrit and two youtube videos to see what kind of to-do I was going for. I found one i like where you don't remove it immediately but crossing it out, but will have to find a way to add localStorage into it.